requirements.txt
flask
lxml
openssh_wrapper
paramiko
passlib
proxmoxer
pyyaml
requests
six
yinsolidated
